[new] Mejoras en la autenticación/autorización + Admin

+++ sd-bank-core-api/src/apis/core/controllers/perfil.controller.ts
@@ -2,25 +2,42 @@ import { Response, Request, NextFunction } from "express";
 import { asyncHandler } from "../../../shared/middlewares/async.middleware";
 import { notFound } from "../../../shared/utils/err.helpers";
 import Perfil from "../../../shared/models/Perfil";
+import { errorHandler } from "../../../shared/middlewares/error.middleware";
+import { ClientSession, startSession } from "mongoose";
+import Usuario from "../../../shared/models/Usuario";
+import ErrorResponse from "../../../shared/utils/error-response";
 
 // @desc   Create perfil
 // @route  POST /api/v1/perfiles
 // @access Private
 export const createPerfil = asyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
-    const { rol, descripcion } = req.body;
+    const session: ClientSession = await startSession();
 
-    const perfilACrear = {
-      rol,
-      descripcion,
-    };
+    try {
+      session.startTransaction();
+      const { rol, descripcion } = req.body;
 
-    await Perfil.create(perfilACrear);
+      const perfilACrear = {
+        rol,
+        descripcion,
+      };
 
-    res.status(201).json({
-      exito: true,
-      mensaje: "¡Perfil agregado satisfactoriamente!",
-    });
+      await Perfil.create(perfilACrear);
+
+      await session.commitTransaction();
+      session.endSession();
+
+      res.status(201).json({
+        exito: true,
+        mensaje: "¡Perfil agregado satisfactoriamente!",
+      });
+    } catch (error) {
+      await session.abortTransaction();
+      session.endSession();
+
+      return errorHandler(error, req, res, next);
+    }
   }
 );
 
@@ -55,21 +72,34 @@ export const getPerfilPorId = asyncHandler(
 // @access Private
 export const updatePerfil = asyncHandler(
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-    const { _id } = req.params;
-    const { descripcion, rol } = req.body;
+    const session: ClientSession = await startSession();
+    session.startTransaction();
 
-    const perfilEncontrado = await Perfil.findById(_id);
+    try {
+      const { _id } = req.params;
+      const { descripcion, rol } = req.body;
 
-    if (!perfilEncontrado) return notFound({ entity: "Perfil", next });
+      const perfilEncontrado = await Perfil.findById(_id);
 
-    const perfilActualizado = { descripcion, rol };
+      if (!perfilEncontrado) return notFound({ entity: "Perfil", next });
 
-    await Perfil.updateOne(perfilEncontrado, perfilActualizado);
+      const perfilActualizado = { descripcion, rol };
 
-    res.status(200).json({
-      exito: true,
-      mensaje: "¡Perfil actualizado satisfactoriamente!",
-    });
+      await Perfil.updateOne(perfilEncontrado, perfilActualizado);
+
+      await session.commitTransaction();
+      session.endSession();
+
+      res.status(200).json({
+        exito: true,
+        mensaje: "¡Perfil actualizado satisfactoriamente!",
+      });
+    } catch (error) {
+      await session.abortTransaction();
+      session.endSession();
+
+      return errorHandler(error, req, res, next);
+    }
   }
 );
 
@@ -78,17 +108,42 @@ export const updatePerfil = asyncHandler(
 // @access Private
 export const deletePerfil = asyncHandler(
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-    const { _id } = req.params;
+    const session: ClientSession = await startSession();
+    session.startTransaction();
 
-    const perfilEncontrado = await Perfil.findById(_id);
+    try {
+      const { _id } = req.params;
 
-    if (!perfilEncontrado) return notFound({ entity: "Perfil", next });
+      const perfilEncontrado = await Perfil.findById(_id);
 
-    await Perfil.deleteOne(perfilEncontrado);
+      if (!perfilEncontrado) return notFound({ entity: "Perfil", next });
 
-    res.status(200).json({
-      exito: true,
-      mensaje: "¡Perfil eliminado satisfactoriamente!",
-    });
+      const usuarioAsociado: any = Usuario.findOne({
+        perfil: perfilEncontrado._id,
+      });
+
+      if (usuarioAsociado)
+        return next(
+          new ErrorResponse(
+            "No puede eliminar este perfil debido a que está asociado a un usuario.",
+            400
+          )
+        );
+
+      await Perfil.deleteOne(perfilEncontrado);
+
+      await session.commitTransaction();
+      session.endSession();
+
+      res.status(200).json({
+        exito: true,
+        mensaje: "¡Perfil eliminado satisfactoriamente!",
+      });
+    } catch (error) {
+      await session.abortTransaction();
+      session.endSession();
+
+      return errorHandler(error, req, res, next);
+    }
   }
 );
